refactor: 💡 renaming, code seperation and missing props
